+++ CHANGELOG.md
@@ -18,7 +18,9 @@ Each version section may have subsections for: _Added_, _Changed_, _Removed_, _D
 - Added support for training with multiple prepared data directories: `sockeye-train --prepared-data dir1 dir2 ...`.
   - Prepared data sources should all use the same vocabularies (specify `sockeye-prepare-data --source-vocab ... sockeye-prepare-data --target-vocab ...`).
   - At each training step, one of the prepared data sources is randomly chosen to provide the next batch.
-  - By default, all data sources have an equal chance of being chosen (`--data-sampling-method uniform`). Alternatively, the choice can be weighted by data size (number of sequences) and a temperature parameter (`--data-sampling-method temperature --data-sampling-temperature ...`) as described by [Arivazhagan et al (2019)](https://aclanthology.org/N19-1388/).
+  - By default, all data sources have an equal chance of being chosen (`--data-sampling-method uniform`).
+  - Alternatively, the choice can be weighted by data size (number of sequences) and a temperature parameter (`--data-sampling-method temperature --data-sampling-temperature ...`) as described by [Arivazhagan et al (2019)](https://aclanthology.org/N19-1388/).
+  - The user can also specify custom weights (`--data-sampling-method custom --data-sampling-custom ...`).
 
 ## [3.0.11]